ML-model/requirements.txt
gradio>=4.44.0
torch>=2.0.0
torchcrf>=1.1.0
openai-whisper>=20231117
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
